Theo

+++ src/main/java/com/Aster/Model/Cart.java
@@ -6,10 +6,10 @@ import java.util.*;
 public class Cart {
     //private List<Order> cart = new ArrayList<>();
     private List<Order> cart;
-    private int totalprice;
+    private double totalprice;
 
     public Cart(@JsonProperty("cart") List<Order> cart,
-                @JsonProperty("totalprice") int totalprice){
+                @JsonProperty("totalprice") double totalprice){
         this.cart = cart;
         this.totalprice = totalprice;
     }
@@ -26,11 +26,11 @@ public class Cart {
         this.cart = cart;
     }
 
-    public int getTotalprice() {
+    public double getTotalprice() {
         return totalprice;
     }
 
-    public void setTotalprice(int totalprice) {
+    public void setTotalprice(double totalprice) {
         this.totalprice = totalprice;
     }
 }
